backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiosqlite
python-jose[cryptography]
passlib
bcrypt
chromadb
sentence-transformers
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf2
python-docx
python-pptx
openpyxl
Pillow
python-dotenv
pydantic[email]
